requirements.txt
discord.py>=2.2.0 
python-whois>=0.8.0
python3-nmap>=1.6.0